+++ StripePaymentsUI/StripePaymentsUI/Source/Helpers/CardElementConfigService.swift
@@ -9,10 +9,9 @@ import Foundation
 
 private let CardElementConfigEndpoint = URL(string: "https://merchant-ui-api.stripe.com/elements/mobile-card-element-config")!
 
-// TODO: Write unit tests with mocked STPAPIClient data
 class CardElementConfigService {
     // The card element does not currently support non-singleton API clients, use the shared one for now.
-    let apiClient = STPAPIClient.shared
+    var apiClient = STPAPIClient.shared
 
     static let shared = CardElementConfigService()