+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### PaymentsUI
 * [Fixed] An issue with `STPPaymentCardTextField`, where the card params were not updated after deleting an empty sub field.
 * [Fixed] Switched to Asset Catalogs and updated to the latest card brand logos.
+* [Added] Adds support for Card Brand Choice in the card text field.
 
 ### Payments
 * [Added] Support for MobilePay bindings.